graphics library has no more clippy warnings

+++ Libraries/graphics/vertex_buffer.rs
@@ -89,24 +89,10 @@ impl VertexBuffer {
             gl::BindBuffer(gl::ARRAY_BUFFER, self.vertex_buffer);
             gl::BindBuffer(gl::ELEMENT_ARRAY_BUFFER, self.index_buffer);
 
-            gl::VertexAttribPointer(0, 2, gl::FLOAT, gl::FALSE, (8 * 4) as i32, std::ptr::null());
-            gl::VertexAttribPointer(
-                1,
-                4,
-                gl::FLOAT,
-                gl::FALSE,
-                (8 * 4) as i32,
-                (2 * 4) as *const _,
-            );
+            gl::VertexAttribPointer(0, 2, gl::FLOAT, gl::FALSE, 8 * 4, std::ptr::null());
+            gl::VertexAttribPointer(1, 4, gl::FLOAT, gl::FALSE, 8 * 4, (2 * 4) as *const _);
             if has_texture {
-                gl::VertexAttribPointer(
-                    2,
-                    2,
-                    gl::FLOAT,
-                    gl::FALSE,
-                    (8 * 4) as i32,
-                    (6 * 4) as *const _,
-                );
+                gl::VertexAttribPointer(2, 2, gl::FLOAT, gl::FALSE, 8 * 4, (6 * 4) as *const _);
             }
 
             gl::EnableVertexAttribArray(0);
